requirements.txt
openai-whisper>=20231117
argostranslate>=1.9.0
